requirements.txt
# ── Data Fetching ────────────────────────────────────────────────────────────
tiingo==0.15.1
requests==2.32.3

# ── Data Processing ──────────────────────────────────────────────────────────
pandas==2.2.2
pyarrow==16.1.0
db-dtypes==1.3.0

# ── Google Cloud ─────────────────────────────────────────────────────────────
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0
google-auth==2.28.1
google-api-core==2.19.0

# ── Environment & Config ─────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Type Checking (dev) ───────────────────────────────────────────────────────
mypy-extensions==1.0.0
